requirements.txt

flask
pdf2image
pillow
opencv-python
fpdf
